Rename _schema.py file to schema
