+++ vantage6-server/vantage6/server/__init__.py
@@ -43,7 +43,7 @@ from vantage6.common.globals import PING_INTERVAL_SECONDS
 from vantage6.server import db
 from vantage6.cli.context import ServerContext
 from vantage6.server.model.base import DatabaseSessionManager, Database
-from vantage6.server.resource.common._schema import HATEOASModelSchema
+from vantage6.server.resource.common.schema import HATEOASModelSchema
 from vantage6.server.permission import RuleNeed, PermissionManager
 from vantage6.server.globals import (
     APPNAME,
