+++ vantage6-server/vantage6/server/resource/__init__.py
@@ -16,7 +16,7 @@ from flask_socketio import SocketIO
 
 from vantage6.common import logger_name
 from vantage6.server import db
-from vantage6.server.resource.common._schema import HATEOASModelSchema
+from vantage6.server.resource.common.schema import HATEOASModelSchema
 from vantage6.server.permission import PermissionManager
 from vantage6.server.resource.common.pagination import Page
