+++ vantage6-server/vantage6/server/resource/collaboration.py
@@ -12,7 +12,7 @@ from vantage6.server.permission import (
     Operation as P,
     PermissionManager
 )
-from vantage6.server.resource.common._schema import (
+from vantage6.server.resource.common.schema import (
     CollaborationSchema,
     TaskSchema,
     OrganizationSchema,
