+++ vantage6-server/vantage6/server/resource/role.py
@@ -17,7 +17,7 @@ from vantage6.server.permission import (
     PermissionManager
 )
 from vantage6.server.model.rule import Operation, Scope
-from vantage6.server.resource.common._schema import RoleSchema, RuleSchema
+from vantage6.server.resource.common.schema import RoleSchema, RuleSchema
 from vantage6.server.resource.common.pagination import Pagination
 from vantage6.server.default_roles import DefaultRole
