+++ vantage6-server/vantage6/server/resource/rule.py
@@ -13,7 +13,7 @@ from vantage6.server.resource import (
 )
 from vantage6.common import logger_name
 from vantage6.server import db
-from vantage6.server.resource.common._schema import RuleSchema
+from vantage6.server.resource.common.schema import RuleSchema
 from vantage6.server.resource.common.pagination import Pagination
