+++ vantage6-server/vantage6/server/resource/run.py
@@ -21,7 +21,7 @@ from vantage6.server.resource import (
     ServicesResources
 )
 from vantage6.server.resource.common.pagination import Pagination
-from vantage6.server.resource.common._schema import (
+from vantage6.server.resource.common.schema import (
     RunSchema, RunTaskIncludedSchema, ResultSchema
 )
 from vantage6.server.model import (
